Added

+++ commands/calcularDiv.js
@@ -1,6 +1,6 @@
 module.exports = (client, message, args) => {
     
-    { let numArray = args.map(n=> parseFloat(n));  let total = numArray.reduce( (p,c) => p/c>>0); message.channel.sendMessage(total); }
+    { let numArray = args.map(n=> parseInt(n));  let total = numArray.reduce( (p,c) => p/c>>0); message.channel.sendMessage(total); }
